# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FieldSure API
  description: |
    API for **FieldSure — Digital Companion for Field Drug Testing**.

    FieldSure captures colorimetric field-test images, validates quality,
    calibrates against a reference card, runs a configuration-driven
    classification pipeline, and produces cryptographically signed
    evidence records with full chain-of-custody metadata.

    > **Scientific Limitation:** No scientifically validated model is
    > currently available. The CV/ML infrastructure is implemented and
    > ready for validated, kit-specific labeled data. The current
    > classification configuration identifier is `DEMO-CONFIG-v1`.
  version: 0.1.0
  contact:
    name: FieldSure Development Team
servers:
  - url: http://localhost:3000
    description: Local development server

# ============================================================
# Security
# ============================================================
security: []   # default — each endpoint declares its own
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token issued by `POST /api/v1/auth/login`.
        Include as `Authorization: Bearer <token>`.
        Tokens expire (default 15 m) and can be refreshed via
        `POST /api/v1/auth/refresh`.

  # ============================================================
  # Reusable Schemas
  # ============================================================
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
      required: [timestamp]

    # ---------- Error ----------
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              example: 400
            message:
              type: string
              example: Validation failed
            details:
              type: array
              items:
                type: string
              description: Present only for validation errors (class-validator).
          required: [code, message]
        meta:
          $ref: '#/components/schemas/Meta'
      required: [error, meta]

    # ---------- Auth ----------
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        operatorId:
          type: string
          example: BADGE-001
        name:
          type: string
          example: Officer Demo
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, operatorId, name, email, role, isActive, createdAt]
...
    UserRole:
      type: string
      enum: [ADMIN, SUPERVISOR, OPERATOR]
    CreateUserRequest:
      type: object
      properties:
        operatorId:
          type: string
          example: BADGE-042
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/UserRole'
      required: [operatorId, name, email, password]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
        operatorId:
          type: string

    # ---------- Test Kits ----------
...
paths:
...
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      security:
        - bearerAuth: []
      description: "**Requires role:** ADMIN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email or operatorId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Users]
      summary: List all users
      operationId: listUsers
      security:
        - bearerAuth: []
      description: "**Requires role:** ADMIN or SUPERVISOR"
      responses:
        '200':
          description: User list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUser
      security:
        - bearerAuth: []
      description: "**Requires role:** ADMIN or SUPERVISOR"
      responses:
        '200':
          description: User details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      security:
        - bearerAuth: []
      description: "**Requires role:** ADMIN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email or operatorId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Users]
      summary: Deactivate user
      operationId: deactivateUser
      security:
        - bearerAuth: []
      description: |
        **Requires role:** ADMIN

        Soft-deletes the user by setting `isActive = false` and
        revokes all active auth sessions.
      responses:
        '200':
          description: User deactivated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: User deactivated
                  meta:
                    $ref: '#/components/schemas/Meta'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================================
  # Test Kits
  # ==========================================================
...